pyproject.toml
[tool.poetry]
name = "restframework-action-permission"
version = "0.2.0"
description = "Action based way of handling permissions in DRF."
authors = ["Hadi Mohseni <hadimohseni1996@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/Hadi-Mohseni/restframework-action-permission"
keywords = ["Django", "Django rest framework", "Django restframework permission"]




[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.3"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"